Missile Sprite on PlayerUI with scripts and Crosshair Change
I added a sprite for the missiles that replaced the bar and counter and added some code in the player function script for the missile to change color, but I can't test it since I don't have a xbox controller. Also changed the crosshair to be less view blocking for the player and it's kinda transparent too.